server/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
alembic==1.13.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
requests==2.31.0
pillow==10.2.0
pandas==2.1.4 